Create TH112xZB_Sinope_Hubitat.groovy
This driver in comparison to the default driver given by Hubitat for this device allows to shows the outside temperature by sending this value by notification. This device also have an energy meter, this driver also retrieved this value. For odd reason, sometimes the energy meter gives lower value than the precedent, so an offset for this value is kept in the state variable, so the value always increase as it should.